requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
paho-mqtt==1.6.1
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
python-socketio==5.11.2
eventlet==0.36.1